Added the answers section
In this section, all answers will be displayed with their respective
user information in a vertical tiled format. Once the database is
connected, the formatting will take care of itself based on the current
css implementation. Finally, a form for submitting an answer was added
at the bottom of the user answers section with a limit of 1000 words
(for now)