Updates on main app.js
Code implementation of forum page get function line 161. Two nested
queries are created to get values from question and answer tables from
the database; an object outputQ is created where its parameters are
defined from the database. Right now it pulls the last element on the
question table

On same object, an answer parameter is created as an array that
dynamically displays all the answers related to the question_id.