Updates on main app.js file
Code implementation for querying a new question from user to the
database. A new object newQ  has its parameters parsed form the ask
question form, then the object is queried using sql2 to the question
table of the database. After a new question is successfully querried
the webpage redirects the user to a new forum page of his/her answer