Updated ask_page.js
Modification of the last query, sql3.

Added a connection of question  with user table, where the query gets
all the question ids from the question table that are from the user id
specified (in our case the user who asked the question)

Lastly, after success on posting a question, the system redirects the
user to a forum page of his/her new question asked.