Added Signup Existence Check
The back-end now checks if the desired sign-up username OR email already
exists. If it does, the user is warned with a text-based message on the
sign-up page, and no account is added to the database. If the username
AND email are both unique, then the account is logged in the "user"
database table. #198  Must still add a re-direct to login page if the
signup is a valid one.