Added compelte sign-up feature
See #198 for full description. The user is now given the sign-up
attempt, and if either their username or email is already taken, the
attempt will be rejected and they will be warned via an embedded
message. If the attempt is successful, the user will be notified via a
redirect to confirmation page. Here they can click on a "Click here to
login" and get to using the site.