Added Cookie at Login
When the user lands at the login page, javascript checks for cookie
compatability on the client-side and locks the login button from
submission if cookies are turned off. If the cookies are turned on and
the user enters the correct email and password, a cookie will be stored
client-side with the username and user_id, plus an extra boolean called
"logged" to allow for quick login checks. Middleware was also added that
sends the user cookie (if it exists) to response object, allowing ejs
files to access it.