Added Cookie Support
The user can now log in to the system if they have cookies enabled in
their browser. They may navigate with a client-side cookie that
maintains the relevant data used to fill certain pages and forms. The
actual filling of the forms and so on should be done in another issue.
see #196