Updated the vote.js
I added the 3rd query to both the answer post and the question post and made the necessary changes to "res.send" the total score for the specific question or answer.
Currently its non-functional. But you should be able to understand how it works based on the comments.
The answer post and the question post are practically the same.