Fixed Display of total votes of answers
Updated forum_page.js to implement the display of total amount of votes
for all the answer in a question_page. There is a query from
score_answer table form our database and for each answer_id the system
calculates the sum of all the scores in it. Going through  several
scenarios, the final score vote to be displayed is selected .

Also modifications to forum_page.ejs to actually display a value from
forum_page.js