Added Favorite Button to Forum Page
The forum page now has a star that allows the user who asked the
question to choose a favorite question. The format is:

if the user is logged in and asked the question, the star is visible,
and the id will be "fav-<answer_id>" else it will be
"fav-null-<answer_id>". One can test if the button should be linked to a
click event listener by grabbing all the stars with the class
"favorite-placeholder". Each of these id's should be test with
element.id.split("-")[1] !== "null"