Added functionality to select best answer
The user who asked a certain question now is able to click on a star to
choose the best answer. The backend aspect is not complete, but will be
added in further commits. The backend should send the front-end a "+" in
the case where the favorite was appropriately saved, else send a "-"