Added delete functionality
The user is now able to delete the question he asked and the answers he provided.

The delete functionality will work only if logged in user match the user_id cookie. This makes sure that a user cannot delete a question or answer of another user. A HTTP GET method is used to get the question_id or answer_id from the delete/submit button from the userprofile page. The queries will execute and the user will be redirected to the same page.

A modal is used for the user to confirm his decision.

The database has been modified to allow CASCADE delete, as such, deleting a question will also delete all the answers from that question. The entire row is deleted.